Split "formatted" types into own file.
